+++ Woof/WoofTests/Model/Preferences/PreferencesHandlerSaveTests.swift
@@ -1,0 +1,29 @@
+import XCTest
+
+final class PreferencesHandlerSaveTests: XCTestCase {
+    private var userPreferencesStorage = KeyValueStorage.preferencesStorage
+
+    override func setUp() {
+        super.setUp()
+        userPreferencesStorage.removeAll()
+    }
+
+    override func tearDown() {
+        userPreferencesStorage.removeAll()
+        super.tearDown()
+    }
+
+    func testSaveSelectedRoleExistsInAPI() {
+        // Given
+        // When
+        // Then
+        _ = PreferencesHandler.set(userRole: .sitter)
+    }
+
+    func testSaveSelectedRoleReturnsTrueForPossibleRoleCase() {
+        // Given
+        // When
+        // Then
+        XCTAssertTrue(PreferencesHandler.set(userRole: .owner))
+    }
+}